fix action class duplication
