+++ src/Services/ApplicationModel/CopyBatchAction.cs
@@ -1,13 +1,14 @@
-﻿using SenseNet.ContentRepository.i18n;
-using System;
-using SenseNet.ContentRepository;
+﻿using System;
 using System.Collections.Generic;
+using System.Linq;
+using SenseNet.ApplicationModel;
+using SenseNet.ContentRepository;
+using SenseNet.ContentRepository.i18n;
 using SenseNet.ContentRepository.Storage;
 using SenseNet.Diagnostics;
-using System.Linq;
 using SenseNet.Portal.OData;
 
-namespace SenseNet.ApplicationModel
+namespace SenseNet.Services.ApplicationModel
 {
     public class CopyBatchAction : CopyToAction
     {
@@ -31,7 +32,7 @@ namespace SenseNet.ApplicationModel
 
         public override bool IsODataOperation { get { return true; } }
 
-        public override ActionParameter[] ActionParameters { get; } = 
+        public override ActionParameter[] ActionParameters { get; } =
         {
             new ActionParameter("targetPath", typeof (string), true),
             new ActionParameter("paths", typeof (object[]), true)
@@ -40,37 +41,62 @@ namespace SenseNet.ApplicationModel
         public override object Execute(Content content, params object[] parameters)
         {
             var targetPath = (string)parameters[0];
-            var exceptions = new List<Exception>();
-
             var targetNode = Node.LoadNode(targetPath);
             if (targetNode == null)
                 throw new ContentNotFoundException(targetPath);
 
             var ids = parameters[1] as object[];
             if (ids == null)
-                throw new InvalidOperationException("No content identifiers provided.");
-            
-            foreach (var node in Node.LoadNodes(ids.Select(NodeIdentifier.Get)))
             {
+                throw new InvalidOperationException("No content identifiers provided.");
+            }
+            var results = new List<object>();
+            var errors = new List<ErrorContent>();
+            var identifiers = ids.Select(NodeIdentifier.Get).ToList();
+            var nodes = Node.LoadNodes(identifiers).ToList();
+            foreach (var id in identifiers)
+            {
+                var node = nodes.FirstOrDefault(n => n.Id == id.Id || n.Path == id.Path);
+                if (node == null)
+                {
+                    errors.Add(new ErrorContent
+                    {
+                        Content = new { id.Id, id.Path, Name = "" },
+                        Error = new Error
+                        {
+                            Code = "ResourceNotFound",
+                            ExceptionType = null
+                            ,InnerError = null
+                            ,Message = new ErrorMessage { Lang = System.Globalization.CultureInfo.CurrentUICulture.Name.ToLower(), Value = null }
+                        }
+                    });
+                    continue;
+                }
                 try
                 {
-                    node?.CopyTo(targetNode);
+                    var copy = node.CopyToAndGetCopy(targetNode);
+                    results.Add(new { copy.Id, copy.Path, copy.Name });
                 }
                 catch (Exception e)
                 {
-                    exceptions.Add(e);
-
                     //TODO: we should log only relevant exceptions here and skip
                     // business logic-related errors, e.g. lack of permissions or
                     // existing target content path.
                     SnLog.WriteException(e);
+                    errors.Add(new ErrorContent
+                    {
+                        Content = new { node.Id, node.Path, node.Name },
+                        Error = new Error
+                        {
+                            Code = "NotSpecified",
+                            ExceptionType = e.GetType().FullName
+                            ,InnerError = new StackInfo { Trace = e.StackTrace }
+                            ,Message = new ErrorMessage { Lang = System.Globalization.CultureInfo.CurrentUICulture.Name.ToLower(), Value = e.Message }
+                        }
+                    });
                 }
             }
-
-            if (exceptions.Count > 0)
-                throw new Exception(string.Join(Environment.NewLine, exceptions.Select(e => e.Message)));
-
-            return null;
+            return ODataActionResponse.Create(results, errors, results.Count + errors.Count);
         }
     }
 }
